backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
xlsxwriter==3.1.9
google-generativeai==0.3.2
Pillow==10.1.0
gspread==6.0.2
google-auth==2.23.4
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8